Fix: DB Schema Update
Change table name
