+++ src/app/v1/user/entity/location.py
@@ -10,3 +10,6 @@ class Location(BaseModel, models.Model):
 
     def __str__(self):
         return self.name
+
+    class Meta:
+        table = "location"
