+++ src/app/v1/user/entity/screen_info.py
@@ -2,7 +2,7 @@ from tortoise import fields, models
 from common.models.base_model import BaseModel
 
 
-class SreenInfo(BaseModel, models.Model):
+class ScreenInfo(BaseModel, models.Model):
     screen_info_id = fields.IntField(pk=True)
     start_time = fields.DatetimeField()
     end_time = fields.DatetimeField()
@@ -12,3 +12,9 @@ class SreenInfo(BaseModel, models.Model):
     location_id = fields.ForeignKeyField(
         "models.Movie", related_name="Sreen_Info", on_delete=fields.CASCADE
     )
+
+    def __str__(self) -> str:
+        return f"ScreenInfo {self.screen_info_id}: {self.screen_id} "
+
+    class Meta:
+        table = "screeninfo"
