--- src/common/models/base_models/BaseModel.py
@@ -1,9 +1,0 @@
-from tortoise import fields, models
-
-
-class BaseModel(models.Model):
-    created_at = fields.DatetimeField(auto_now_add=True)
-    updated_at = fields.DatetimeField(auto_now=True)
-
-    class Meta:
-        abstract = True
